chore(signals): parameterize and dedupe safety-filter telemetry tests
Address review feedback: collapse the blocked/safe capture cases into one
parameterized test, and move the shared `aorganization`/`ateam` async fixtures
into the existing test-suite `conftest.py` instead of duplicating them across
the signals telemetry test modules.

Generated-By: PostHog Code
Task-Id: 816a9875-f57e-4a6d-a197-fddcd8466c36
